Moving transition to css instead of spring
