+++ lib/src/modules/statusIndicatorCircle/statusIndicatorCircle.js
@@ -1,29 +1,23 @@
 import React from 'react';
-import { useSpring, animated } from 'react-spring';
-import * as easings from 'd3-ease';
+import cx from 'classnames';
 import Icon from '../../../Icon';
 
 export function StatusIndicatorCircle({ color, icon, solid, thickBorder }) {
-  const styles = useSpring({
-    borderWidth: thickBorder ? '4px' : '2px',
-    config: {
-      easing: easings.easeCubic,
-      duration: 300
-    }
+  const className = cx('status-indicator-circle', {
+    'thick-border': thickBorder,
+    'thin-border': !thickBorder
   });
 
   const addedStyles = {
-    ...styles,
     backgroundColor: solid ? color : 'transparent',
-    borderColor: color,
-    borderStyle: 'solid'
+    borderColor: color
   };
 
   return (
-    <animated.span style={addedStyles} className="status-indicator-circle">
+    <span style={addedStyles} className={className}>
       {icon && (
         <Icon name={icon} types={['white']} defaultActiveColor={false} />
       )}
-    </animated.span>
+    </span>
   );
 }
