+++ lib/src/modules/statusIndicatorCircle/statusIndicatorCircleStory.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import StoryItem from 'stories/styleguide/StoryItem';
 import { withKnobs, color, boolean } from '@storybook/addon-knobs';
-import { StatusIndicatorCircle } from './statusIndicator';
+import { StatusIndicatorCircle } from './statusIndicatorCircle';
 
 const stories = storiesOf('Modules', module);
 stories.addDecorator(withKnobs);
